32-bit compatability

+++ src/ERFA.jl
@@ -262,7 +262,7 @@ export
 include("../deps/deps.jl")
 include("erfa_common.jl")
 
-function getindex(A::Array_3_Cdouble, i::Int64)
+function getindex(A::Array_3_Cdouble, i::Integer)
     if i === 1
         return A.d1
     elseif i === 2
@@ -272,7 +272,7 @@ function getindex(A::Array_3_Cdouble, i::Int64)
     end
 end
 
-function getindex(A::Array_3_Array_3_Cdouble, i::Int64)
+function getindex(A::Array_3_Array_3_Cdouble, i::Integer)
     if i === 1
         return getindex(A.d1,1)
     elseif i === 2
@@ -949,7 +949,7 @@ function eraPfw06(date1::Cdouble,date2::Cdouble)
     gamb[1],phib[1],psib[1],epsa[1]
 end
 
-function eraPlan94(date1::Float64, date2::Float64, np::Int64)
+function eraPlan94(date1::Float64, date2::Float64, np::Integer)
     pv = zeros(6)
     i = ccall((:eraPlan94, liberfa),Cint,
               (Float64, Float64, Int64, Ptr{Float64}),
@@ -1283,7 +1283,7 @@ for f in (:eraA2af,
           :eraA2tf,
           :eraD2tf)
     @eval begin
-        function ($f)(ndp::Int64, a::Float64)
+        function ($f)(ndp::Integer, a::Float64)
             s = "+"
             i = Int32[0, 0, 0, 0]
             ccall(($(Expr(:quote,f)),liberfa),Void,
